--- cmd/rpcdaemon/.gitignore
@@ -1,1 +1,0 @@
-*.http
