+++ cmd/rpcdaemon/Dockerfile
@@ -11,10 +11,10 @@ ADD . /app
 RUN cd /app && make rpcdaemon
 
 # Pull Geth into a second stage deploy alpine container
-FROM alpine:latest
+FROM alpine:3
 
 RUN apk add --no-cache ca-certificates
 COPY --from=builder /app/build/bin/rpcdaemon /usr/local/bin/
 
-EXPOSE 8545 30303 30303/udp 9090 6060
+EXPOSE 8545
 ENTRYPOINT ["rpcdaemon"]
