+++ cmd/rpcdaemon/commands/daemon.go
@@ -2,20 +2,15 @@ package commands
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"math/big"
 	"strings"
 
 	"github.com/ledgerwatch/turbo-geth/common"
-	"github.com/ledgerwatch/turbo-geth/common/hexutil"
 	"github.com/ledgerwatch/turbo-geth/consensus"
-	"github.com/ledgerwatch/turbo-geth/core"
 	"github.com/ledgerwatch/turbo-geth/core/rawdb"
 	"github.com/ledgerwatch/turbo-geth/core/state"
 	"github.com/ledgerwatch/turbo-geth/core/types"
-	"github.com/ledgerwatch/turbo-geth/core/vm"
-	"github.com/ledgerwatch/turbo-geth/eth/stagedsync/stages"
 	"github.com/ledgerwatch/turbo-geth/ethdb"
 	"github.com/ledgerwatch/turbo-geth/internal/ethapi"
 	"github.com/ledgerwatch/turbo-geth/log"
@@ -33,176 +28,6 @@ func splitAndTrim(input string) []string {
 		result[i] = strings.TrimSpace(r)
 	}
 	return result
-}
-
-// EthAPI is a collection of functions that are exposed in the
-type EthAPI interface {
-	Coinbase(context.Context) (common.Address, error)
-	BlockNumber(ctx context.Context) (hexutil.Uint64, error)
-	GetBlockByNumber(ctx context.Context, number rpc.BlockNumber, fullTx bool) (map[string]interface{}, error)
-	GetBalance(_ context.Context, address common.Address, blockNrOrHash rpc.BlockNumberOrHash) (*hexutil.Big, error)
-	GetTransactionReceipt(ctx context.Context, hash common.Hash) (map[string]interface{}, error)
-	GetLogs(ctx context.Context, hash common.Hash) ([][]*types.Log, error)
-	Call(ctx context.Context, args ethapi.CallArgs, blockNrOrHash rpc.BlockNumberOrHash, overrides map[common.Address]ethapi.Account) (hexutil.Bytes, error)
-	EstimateGas(ctx context.Context, args ethapi.CallArgs) (hexutil.Uint64, error)
-}
-
-// APIImpl is implementation of the EthAPI interface based on remote Db access
-type APIImpl struct {
-	db           ethdb.KV
-	dbReader     ethdb.Getter
-	chainContext core.ChainContext
-}
-
-// NewAPI returns APIImpl instance
-func NewAPI(db ethdb.KV, dbReader ethdb.Getter, chainContext core.ChainContext) *APIImpl {
-	return &APIImpl{
-		db:           db,
-		dbReader:     dbReader,
-		chainContext: chainContext,
-	}
-}
-
-func (api *APIImpl) BlockNumber(ctx context.Context) (hexutil.Uint64, error) {
-	execution, _, err := stages.GetStageProgress(api.dbReader, stages.Execution)
-	if err != nil {
-		return 0, err
-	}
-	return hexutil.Uint64(execution), nil
-}
-
-// Etherbase is the address that mining rewards will be send to
-func (api *APIImpl) Coinbase(_ context.Context) (common.Address, error) {
-	return common.Address{}, errors.New("not implemented")
-}
-
-func GetReceipts(ctx context.Context, db rawdb.DatabaseReader, cfg *params.ChainConfig, hash common.Hash) (types.Receipts, error) {
-	number := rawdb.ReadHeaderNumber(db, hash)
-	if number == nil {
-		return nil, fmt.Errorf("block not found: %x", hash)
-	}
-
-	block := rawdb.ReadBlock(db, hash, *number)
-	if cached := rawdb.ReadReceipts(db, block.Hash(), block.NumberU64(), cfg); cached != nil {
-		return cached, nil
-	}
-
-	bc := &blockGetter{db}
-	_, _, ibs, dbstate, err := ComputeTxEnv(ctx, bc, params.MainnetChainConfig, &chainContext{db: db}, db.(ethdb.HasKV).KV(), hash, 0, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	var receipts types.Receipts
-	cc := &chainContext{db}
-	gp := new(core.GasPool).AddGas(block.GasLimit())
-	var usedGas = new(uint64)
-	for i, tx := range block.Transactions() {
-		ibs.Prepare(tx.Hash(), block.Hash(), i)
-
-		header := rawdb.ReadHeader(db, hash, *number)
-		receipt, err := core.ApplyTransaction(params.MainnetChainConfig, cc, nil, gp, ibs, dbstate, header, tx, usedGas, vm.Config{}, nil)
-		if err != nil {
-			return nil, err
-		}
-		receipts = append(receipts, receipt)
-	}
-
-	return receipts, nil
-}
-
-func (api *APIImpl) GetLogs(ctx context.Context, hash common.Hash) ([][]*types.Log, error) {
-	number := rawdb.ReadHeaderNumber(api.dbReader, hash)
-	if number == nil {
-		return nil, fmt.Errorf("block not found: %x", hash)
-	}
-
-	receipts, err := GetReceipts(ctx, api.dbReader, params.MainnetChainConfig, hash)
-	if err != nil {
-		return nil, fmt.Errorf("getReceipt error: %v", err)
-	}
-	logs := make([][]*types.Log, len(receipts))
-	for i, receipt := range receipts {
-		logs[i] = receipt.Logs
-	}
-	return logs, nil
-}
-
-func (api *APIImpl) GetTransactionReceipt(ctx context.Context, hash common.Hash) (map[string]interface{}, error) {
-	// Retrieve the transaction and assemble its EVM context
-	tx, blockHash, blockNumber, txIndex := rawdb.ReadTransaction(api.dbReader, hash)
-	if tx == nil {
-		return nil, fmt.Errorf("transaction %#x not found", hash)
-	}
-
-	receipts, err := GetReceipts(ctx, api.dbReader, params.MainnetChainConfig, blockHash)
-	if err != nil {
-		return nil, fmt.Errorf("getReceipt error: %v", err)
-	}
-	if len(receipts) <= int(txIndex) {
-		return nil, fmt.Errorf("block has less receipts than expected: %d <= %d, block: %d", len(receipts), int(txIndex), blockNumber)
-	}
-	receipt := receipts[txIndex]
-
-	var signer types.Signer = types.FrontierSigner{}
-	if tx.Protected() {
-		signer = types.NewEIP155Signer(tx.ChainID().ToBig())
-	}
-	from, _ := types.Sender(signer, tx)
-
-	// Fill in the derived information in the logs
-	if receipt.Logs != nil {
-		for i, log := range receipt.Logs {
-			log.BlockNumber = blockNumber
-			log.TxHash = hash
-			log.TxIndex = uint(txIndex)
-			log.BlockHash = blockHash
-			log.Index = uint(i)
-		}
-	}
-
-	// Now reconstruct the bloom filter
-	fields := map[string]interface{}{
-		"blockHash":         blockHash,
-		"blockNumber":       hexutil.Uint64(blockNumber),
-		"transactionHash":   hash,
-		"transactionIndex":  hexutil.Uint64(txIndex),
-		"from":              from,
-		"to":                tx.To(),
-		"gasUsed":           hexutil.Uint64(receipt.GasUsed),
-		"cumulativeGasUsed": hexutil.Uint64(receipt.CumulativeGasUsed),
-		"contractAddress":   nil,
-		"logs":              receipt.Logs,
-		"logsBloom":         receipt.Bloom,
-	}
-
-	// Assign receipt status or post state.
-	if len(receipt.PostState) > 0 {
-		fields["root"] = hexutil.Bytes(receipt.PostState)
-	} else {
-		fields["status"] = hexutil.Uint(receipt.Status)
-	}
-	if receipt.Logs == nil {
-		fields["logs"] = [][]*types.Log{}
-	}
-	// If the ContractAddress is 20 0x0 bytes, assume it is not a contract creation
-	if receipt.ContractAddress != (common.Address{}) {
-		fields["contractAddress"] = receipt.ContractAddress
-	}
-	return fields, nil
-}
-
-type blockGetter struct {
-	dbReader rawdb.DatabaseReader
-}
-
-func (g *blockGetter) GetBlockByHash(hash common.Hash) *types.Block {
-	return rawdb.ReadBlockByHash(g.dbReader, hash)
-
-}
-
-func (g *blockGetter) GetBlock(hash common.Hash, number uint64) *types.Block {
-	return rawdb.ReadBlock(g.dbReader, hash, number)
 }
 
 type chainContext struct {
