+++ cmd/rpcdaemon/test.http
@@ -1,0 +1,66 @@
+
+###
+
+POST localhost:8545
+Content-Type: application/json
+
+{ "jsonrpc": "2.0", "method": "debug_accountRange", "params": ["0x1e8480", "", 256, false, false, false], "id":1}
+
+###
+
+POST localhost:8545
+Content-Type: application/json
+
+{"jsonrpc":"2.0","method":"eth_getBlockByNumber","params":["0xf4240",true],"id":2}
+
+###
+
+POST localhost:8545
+Content-Type: application/json
+
+{
+  "jsonrpc": "2.0",
+  "method": "debug_storageRangeAt",
+  "params": [
+    "0x4ced0bc30041f7f4e11ba9f341b54404770c7695dfdba6bb64b6ffeee2074177",
+    99,
+    "0x33990122638b9132ca29c723bdf037f1a891a70c",
+    "0x0000000000000000000000000000000000000000000000000000000000000000",
+    1024
+  ],
+  "id": 537758
+}
+### > 60
+
+### >20
+###{"jsonrpc":"2.0","method":"debug_storageRangeAt","params":["0x6e6ec30ba20b263d1bdf6d87a0b1b037ea595929ac10ad74f6b7e1890fdad744", 19,"0x793ae8c1b1a160bfc07bfb0d04f85eab1a71f4f2","0x0000000000000000000000000000000000000000000000000000000000000000",1024],"id":113911}
+
+
+### {"jsonrpc":"2.0","mesthod":"debug_storageRangeAt","params":["0xbcb55dcb321899291d10818dd06eaaf939ff87a717ac40850b54c6b56e8936ff", 2,"0xca7c390f8f843a8c3036841fde755e5d0acb97da","0x0000000000000000000000000000000000000000000000000000000000000000",1024],"id":3836}
+
+###{"jsonrpc":"2.0","method":"debug_storageRangeAt","params":["0xf212a7655339852bf58f7e1d66f82256d22d13ccba3068a9c47a635738698c84", 0,"0xb278e4cb20dfbf97e78f27001f6b15288302f4d7","0x0000000000000000000000000000000000000000000000000000000000000000",1024],"id":8970}
+
+###
+
+POST localhost:8545
+Content-Type: application/json
+
+{
+  "jsonrpc": "2.0",
+  "method": "eth_getTransactionReceipt",
+  "params": ["0xc05ce241bec59900356ede868d170bc01d743c3cd5ecb129ca99596593022771"],
+  "id": 537758
+}
+
+
+###
+
+POST localhost:8545
+Content-Type: application/json
+
+{
+  "jsonrpc": "2.0",
+  "method": "eth_getLogs",
+  "params": ["0xcbf9651e8fd1858c9f3ba6ae96a093d2e03ce4ba1e966484b84aec4fc7026d2e"],
+  "id": 537758
+}
